requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
# uvloop==0.21.0  # No compatible con Windows
watchfiles==1.1.0
websockets==15.0.1

# Dependencias para base de datos SQL Server & PostgreSQL (Render)
sqlalchemy==2.0.43
pyodbc==5.2.0
alembic==1.16.5
psycopg[binary,pool]==3.2.3
psycopg2-binary>=2.9

# Dependencias para configuración centralizada (Pydantic v2)
pydantic-settings==2.10.1

# Dependencias para seguridad y autenticación
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# Dependencias para validación y tipos
email-validator==2.1.1

# Dependencias para logging mejorado (opcional)
loguru==0.7.2

# Dependencias para Google Cloud Storage
google-cloud-storage==2.18.2

# Dependencias para manejo de archivos
python-multipart==0.0.9
